Create core.magics.basic according to new API.
